Introduce a resource config generation task
This commit introduces a GraalVM resource configuration file generation task.
By having a proper file to generate the resource file, it allows correctly
modeling what is required to generate this file, by declaring those as inputs
of the task.

In particular, we can tell that the resource configuration file inputs are:

    - the resources directories of the main source set
    - the classes directories, because they main contain resources which are
generated at compile time by annotation processors in the `classes` directories

It's then Gradle's responsibility to trigger the dependent tasks (e.g `classes`)
whenever this task is called.

Then, the task declares an output, the JSON file, which can in turn be used
as an input of the GraalVM image generation task. Currently, the wiring is manual,
because the native image task doesn't support declaring resources configuration
directories (it should).

With those changes, we preserve up-to-date checking, declare properly the inputs
and outputs of each tasks, which also preserves cacheability. We also don't have
to care about what lifecycle task to call to get the file generated: tasks
which need it just need to declare it as an input.