Rework how the configuration path is passed to the native-image process
The previous implementation was buggy, because while it was setting the arguments,
they were in practice overridden later when the native image task was called.

Instead, this commit adds a proper task input to the native image task to set
the configuration directories.

Note that the solution isn't perfect: currently the task is wired so that the
directory passed is directly the output of the JSON generator task. However,
`native-image` mandates that all configuration files are found in that directory,
including, potentially, other resource files like the reflection config.

This means that in theory we should have one more indirection level: a task
which copies all the generated resources (and potentially hand-crafted ones)
into a temporary directory which is then used as an input to `native-image`.