Replace use of shaded JSON library with adhoc code
In order to avoid depending on the shaded Jackson library, this commit introduces
an adhoc JSON writer. The goal of this writer is _solely_ to support the generation
of the resource file for GraalVM. It is certainly not rock solid and therefore
makes assumptions on what the content of the file will be.

This is good enough for now since we have a limited use case, but it should be
eventually replaced with a hopefully lightweight JSON writer.