Generate the GraalVM resource file even without the Micronaut application plugin
This commit adds support for generating the GraalVM resource-config.json file
even if the Micronaut Application plugin isn't applied (e.g if the application
plugin is).

This basically adds some null checks on extensions, but ideally the plugin should
react to different plugins being applied, and in particular the Micronaut
application plugin.

Fixes #223