Use the official GraalVM plugin
This commit reworks the Gradle plugin to use the official GraalVM
native image plugin. It accidentally fixes a number of ordering
issues which existed in the previous setup.

Because this work relies on an unmerged branch of the official
plugin, it also introduces a settings plugin which responsibility
is to checkout a branch of the GraalVM plugin and use it as an
included build (Gradle rules!). Of course, this has to be removed
before this PR gets merged, it's just a matter of getting
feedback from CI.