Fix for incorrect paths in docker image file
The generate docker image file contained paths to files on the host
instead of paths in the image builder. The consequence is that
resources files, and in particular `application.yml` file were
totally ignored.

This is a hotfix for this issue, but this is NOT a final fix: this
is a very dirty hack, which is going to be removed once we migrate
to the official GraalVM plugin.

Fixes #246
Fixes #243