Integrate with Micronaut AOT
This commit introduces a new `io.micronaut.aot` plugin which integrates
with the Micronaut AOT module. If applied, this plugin will configure
the build to:

1. register some `prepareXXXOptimizations` tasks for optimizing apps
for JIT and native
2. register new `optimizedRun` task which executes a Micronaut app
with the result of the AOT optimizations
3. register new `nativeOptimizedRun` task which executes a Micronaut
app natively (with GraalVM) with the result of the native optimizations

The AOT optimization configuration is exposed with an `aot` sub-extension
in the `micronaut` extension. Currently there's no opt-in because all
the tasks which are being created are isolated from the standard tasks:
it just makes the AOT optimizations visible, without interfering with
the standard execution.

Future work may include an option to replace the standard tasks with
the optimized versions, but it will require more work to get there (in
particular, we need to think what it means for test execution).