Lazily configure the Java exec tasks (#294)
This commit fixes the fact that Groovy sources (and probably Kotlin/Scala/...)
were not added to the watch paths. The problem was that the exec task was eagerly
configured. By moving to `configureEach` we're deferring the configuration to
when it's actually being requested by a user. As a consequence we are sure, when
his happens, that all plugins are applied, so the Groovy sources will be visible.

Fixes #292