Add support for building an AOT Docker image
This commit refactors the plugins in order to be able to build more
docker images. In particular, it allows building a docker image for
the optimized code generated by Micronaut AOT.

Under the hood, the Micronaut extension has a sub-extension for
docker images. When an image is added, it will automatically register
a number of helper tasks for that image. An image basically consists
of application code (in the form of a jar) and arbitrary number of
layers.

In the refactor, layers have been made more generic, in the sense
that we now support multiple layers for libraries, resources and even
class files. This should make it easier to build docker images without
having to build a jar file.

It's worth noting, however, that the "native" docker image is very
specific and will _not_ do the expected thing for AOT: it's actually
a docker image which will compile the JIT AOT optimized application
within a docker image. It will NOT be the native optimized AOT
application packaged in a native file. Future work will have to
determine if we want to proceed this way.

Closes #309