Add support for building an AOT optimize native docker image
This fixes the native image which is generated when using the AOT
optimizations: instead of using the JIT optimized jar, it will now
properly compute the native optimizations, then use it in the Docker
container when building the native image.

As a result, the generated docker image is an AOT optimized image
for native.