Integrate with Micronaut AOT (#280)
* Integrate with Micronaut AOT

This commit introduces a new `io.micronaut.aot` plugin which integrates
with the Micronaut AOT module. If applied, this plugin will configure
the build to:

1. register some `prepareXXXOptimizations` tasks for optimizing apps
for JIT and native
2. register new `optimizedRun` task which executes a Micronaut app
with the result of the AOT optimizations
3. register new `nativeOptimizedRun` task which executes a Micronaut
app natively (with GraalVM) with the result of the native optimizations

The AOT optimization configuration is exposed with an `aot` sub-extension
in the `micronaut` extension. Currently there's no opt-in because all
the tasks which are being created are isolated from the standard tasks:
it just makes the AOT optimizations visible, without interfering with
the standard execution.

Future work may include an option to replace the standard tasks with
the optimized versions, but it will require more work to get there (in
particular, we need to think what it means for test execution).

* Add some logging for Windows builds

* Add an optimized distribution

This commit adds a distribution which packages the AOT optimized
application. It will automatically add the following tasks:

- `optimizedDistZip` and variants for archives
- `installOptimizedDist` to install the optimized application

So it's possible to test the output by running:

$ ./gradlew installOptimizedDist
$ build/install/demo-app-optimized/bin/demo-app

* Add group/description to the run task

* Update to latest core naming

* Add dependency to the new standard optimizers

* Align with AOT to default to false for optimizations

Micronaut AOT disables all optimizations by default. This makes
sure that the Gradle plugin does the same.

* Add missing optimizations to the DSL

This also fixes the AOT integration test.

* Ignore test on Windows

This seems to make native-image crash, for some reason.

* Add support for building an AOT Docker image

This commit refactors the plugins in order to be able to build more
docker images. In particular, it allows building a docker image for
the optimized code generated by Micronaut AOT.

Under the hood, the Micronaut extension has a sub-extension for
docker images. When an image is added, it will automatically register
a number of helper tasks for that image. An image basically consists
of application code (in the form of a jar) and arbitrary number of
layers.

In the refactor, layers have been made more generic, in the sense
that we now support multiple layers for libraries, resources and even
class files. This should make it easier to build docker images without
having to build a jar file.

It's worth noting, however, that the "native" docker image is very
specific and will _not_ do the expected thing for AOT: it's actually
a docker image which will compile the JIT AOT optimized application
within a docker image. It will NOT be the native optimized AOT
application packaged in a native file. Future work will have to
determine if we want to proceed this way.

Closes #309

* Add support for building an AOT optimize native docker image

This fixes the native image which is generated when using the AOT
optimizations: instead of using the JIT optimized jar, it will now
properly compute the native optimizations, then use it in the Docker
container when building the native image.

As a result, the generated docker image is an AOT optimized image
for native.

* Add tasks to generate sample AOT configuration

This is useful for documentation purposes, if a user wants to know
about all possible optimizations for a specific Micronaut AOT version
(and possibly modules).

* Fix compilation error after AOT API changes

* Add a task to build an optimized fat jar

This commit adds a new task which will build an optimized fat jar.
This is basically the same as the "shadow" jar except that the
classes it contains are those from the Micronaut AOT optimized
binary.

Fixes #313

* Use release versions of Micronaut Core and Micronaut AOT