Upgrade to native build tools 0.9.8
This commit bumps the dependency of the native build tools plugin
to 0.9.8, which lets us use the "ignoreExistingResourcesConfigFile"
flag for automatic resources detection. This makes it possible to
combine the result of the `micronaut-graal` annotation processor,
which is _incomplete_, with the result of the auto-detection
mechanism (note, however, that the jar would only bundle the result
of the annotation processing phase!).

Fixes #324
Fixes #327
Fixes #328