Make it more convenient to declare AOT plugins
A project like Micronaut Security can contribute new AOT modules.
We need a user-friendly way to declare such extensions, and a way
to configure them.

This commit introduces an `aotPlugins` configuration which makes
it more obvious where to declare the AOT extensions. It also adds
a properties block to the AOT configuration, so that configuration
keys from plugins can be used without resorting to a properties
file.