Introduce a minimal application plugin
Similarly to the minimal library plugin, this plugin configures the
build so that the user can build a Micronaut application, but it
will not apply the Docker, Graalvm or Eclipse plugins. The application
plugin remains as it was before, applying all of them, it's simply
that now there's an option to _only_ build an application.

This commit improves coverage to make sure that the AOT, Docker and
GraalVM plugins work properly independenly of whether the minimal
or full application plugin is applied.