Refactor the plugin architecture (#360)
* Introduce a helper class for plugins

Instead of using the static helpers in a parent type.

* Introduce a minimal library plugin

This commit refactors the Micronaut Library plugin code, so that
we can later properly extract the different aspects of a Micronaut
component and apply each plugin separately (e.g, not systematically
apply the GraalVM or Docker plugin).

To do this, we introduce a "Micronaut Component plugin" which is
responsible for configuring compilation, annotation processing, etc.

Then we have a "minimal" library plugin which applies the Java Library
plugin and this component plugin.

The traditional "io.micronaut.library" plugin does the same as before,
that is to say that it applies the minimal plugin, but also configures
the annotation processing for Eclipse and applies the GraalVM plugin.

Ideally, users should move off this plugin because it does too much.
Instead they should probably apply the minimal plugin, then add the
GraalVM plugin for example (and not the Eclipse one if they don't use
it!).

It's worth noting that at this stage, there's no short id for the
minimal library plugin. This will come after more refactorings.

* Introduce a minimal application plugin

Similarly to the minimal library plugin, this plugin configures the
build so that the user can build a Micronaut application, but it
will not apply the Docker, Graalvm or Eclipse plugins. The application
plugin remains as it was before, applying all of them, it's simply
that now there's an option to _only_ build an application.

This commit improves coverage to make sure that the AOT, Docker and
GraalVM plugins work properly independenly of whether the minimal
or full application plugin is applied.

* Document the new plugins

* Update copyrights