Improve Micronaut AOT configuration
This commit improves integration with Micronaut AOT by exposing 2
new configuration options:

- the "target environments", which let the user explain what
deployment environments will be, letting the analyser run with
those environments set, instead of inferring from the local one
- environment variables: this lets the user declare environment
variables which should be visible when the analysis is performed.

Those changes are required to support deployments to environments
like Oracle Cloud, for which the bean requirements are vastly
different from what you get in a local mode.