Split the build into a multi-project plugins build (#374)
* Introduce a Micronaut Gradle module type

This is the first step in splitting this project into multiple
subprojects. It introduces a Micronaut Gradle plugin type, which
is now applied on the root project, but will be moved to a
subproject (then subprojects).

* Refactor the build into a multi-project build

This commit splits the plugin into multiple subprojects. The goal
is to make sure that if a user applies the minimal plugins, or
that if they use a plugin with optional plugin dependencies (like
the Shadow plugin), then they won't get the optional plugins on
classpath.

It seems that it was broken before, and that we added shadow or
the kotlin plugin on classpath.

Because of the TestKit classloading setup, we have to separate
tests into 2 categories:

- tests in a plugin module itself are restricted to the plugin
only: it _cannot_ have other plugin dependencies (like Shadow,
or another Micronaut plugin like the Micronaut Docker plugin),
otherwise we would put them on the test classpath, which would
incorrectly miss classloading errors
- so tests which require a combination of plugins, for example
a minimal plugin + the docker plugin, or a minimal plugin +
the Shadow plugin, MUST be declared in the new `functional-tests`
project.

This functional tests project makes use of a special setup to
simulate real user builds. For this purposes, it depends on the
publication of the Micronaut Gradle plugins to a local repository,
which is then used in the tests. The current setup to publish
the plugins isn't great but that's all we can do with Gradle for now.

This new setup makes it much easier to detect problems like types
being used in a plugin, which should be available at compile time,
but shouldn't trigger problems if they aren't present at runtime.
This is typically the case if a plugin needs to react to the presence
of the shadow plugin, which is optionally applied by a user.

It's worth noting that for backwards compatibility reasons, the
name `micronaut-gradle-plugin` remains: this plugin corresponds to
the good old Micronaut plugin and applies the other plugins
transparently.

* Fix javadoc generation

* Bump version to 3.3.x

* Fix input tracking of functional tests

* Remove toolchain for tests

Since we're relying on a build matrix to test various versions

* Fix test requirements

* Fix Windows build

* Fix test requirement and remove logging