Rework how the default runtime class is set for Lambda
Before this change, the class used to generate native lambdas was
always set to `MicronautLambdaRuntime`, independently of the fact
that the user declared a `mainClass` or not. With this change, the
main class defined by the user will _always_ be used. Only if no
main class is set, we will default to `MicronautLambdaRuntime`.