Add test resources plugin
This commit introduces a new plugin, `io.micronaut.test-resources`
which integrates with Micronaut Test Resources. It provides support
for starting the test resources server in the background, and lets
the user configure test resources support.

This plugin can be used in combination with other "minimal" plugins:

```
plugins {
     id 'io.micronaut.minimal.application'
     id 'io.micronaut.test-resources'
}
```

It is automatically applied by the `application` and `library`
plugins. Support for test resources is only enabled if the target
Micronaut version is at least 3.5.

The plugin introduces a configuration block:

```
micronaut {
   ...
   testResources {
      // test resources configuration goes here
   }
}
```

By default, the plugin will perform classpath inference
to determine which test resources module to include. There
are several configuration options:

```
testResources {

    enabled = true // by default, only if Micronaut 3.5+

    version = "1.2" // override Test Resources version

    classpathInference = false // true by default

    // Add explicit Micronaut Test resources modules
    // in case inference fails to find them
    additionalModules.addAll("kafka", "hivemq")

    // Instead of using a randomized port, use an
    // explicit port for the server
    explicitPort.set(15455)
}
```

By default, the server will only accept requests from
this build, thanks to a randomly generated access token.
However, if an explicit port is used, token generation
is disabled: this is so that multiple, concurrent builds
can access the same test resources server (which may be
useful when testing interations between a client and
a server in a messaging solution for example).

Last but not least, test resources try to infer dependencies
to be added on the server classpath. There may be cases
where it fails finding dependencies it should add, in
which case you can explicitly add dependencies:

```
dependencies {
    testresources 'some.sql:driver:1.0'
}
```

The server is integrated in tests and development mode.