Rework how test resources are started
This commit makes use of the new test resources build tool
utilities to handle the lifecycle of the test resources server.
There are now multiple options when starting a test server.

Case 1: single build, no existing test server

In this case, a single build is executed, for example `run`
or `test` and nothing special is configured. A test resources
server will be started at the beginning of the build and
automatically shutdown at the end of the build.

Case 2: continuous build, no existing test server

In this case, a test resources server will be spawned at the
beginning of the build session, multiple builds will reuse
the same server (for example when a test changes), and the
test resources server is shutdown at the end of the build
session.

Case 3: project local standalone mode

If the only task which is called is `startTestResourcesService`,
then a test resources server is started in the background and
will _outlive_ the build. There is an option to block the build
instead of running in the background.

When running in the background, the test server must be shutdown
by calling the `stopTestResourcesService` task.

Case 4: single build or continuous build connecting to a
standalone server

If a server is already started in standalone mode, instead of
starting a new test server, builds will reuse the test server
(unless a different _settings directory_ is set).

Case 5: shared test resources server

This is a variant of the standalone mode, where the settings
are written in a project-independent location, making it possible
to share a test resources server between multiple builds, in
multiple projects.