Add test resources plugin (#478)
* Add test resources plugin

This commit introduces a new plugin, `io.micronaut.test-resources`
which integrates with Micronaut Test Resources. It provides support
for starting the test resources server in the background, and lets
the user configure test resources support.

This plugin can be used in combination with other "minimal" plugins:

```
plugins {
     id 'io.micronaut.minimal.application'
     id 'io.micronaut.test-resources'
}
```

It is automatically applied by the `application` and `library`
plugins. Support for test resources is only enabled if the target
Micronaut version is at least 3.5.

The plugin introduces a configuration block:

```
micronaut {
   ...
   testResources {
      // test resources configuration goes here
   }
}
```

By default, the plugin will perform classpath inference
to determine which test resources module to include. There
are several configuration options:

```
testResources {

    enabled = true // by default, only if Micronaut 3.5+

    version = "1.2" // override Test Resources version

    classpathInference = false // true by default

    // Add explicit Micronaut Test resources modules
    // in case inference fails to find them
    additionalModules.addAll("kafka", "hivemq")

    // Instead of using a randomized port, use an
    // explicit port for the server
    explicitPort.set(15455)
}
```

By default, the server will only accept requests from
this build, thanks to a randomly generated access token.
However, if an explicit port is used, token generation
is disabled: this is so that multiple, concurrent builds
can access the same test resources server (which may be
useful when testing interations between a client and
a server in a messaging solution for example).

Last but not least, test resources try to infer dependencies
to be added on the server classpath. There may be cases
where it fails finding dependencies it should add, in
which case you can explicitly add dependencies:

```
dependencies {
    testresources 'some.sql:driver:1.0'
}
```

The server is integrated in tests and development mode.

* Fix `addLater`

For some reason, Gradle complains about the `null`.

* Make Sonar happier

* Ignore test on Windows since Docker is not available

* Introduce constants for known modules

So that users can refer to them instead of strings.

* Do not apply the test resources plugin automatically

This will be handled in starter instead.

* Let the user specify client timeout

* Rework how test resources are started

This commit makes use of the new test resources build tool
utilities to handle the lifecycle of the test resources server.
There are now multiple options when starting a test server.

Case 1: single build, no existing test server

In this case, a single build is executed, for example `run`
or `test` and nothing special is configured. A test resources
server will be started at the beginning of the build and
automatically shutdown at the end of the build.

Case 2: continuous build, no existing test server

In this case, a test resources server will be spawned at the
beginning of the build session, multiple builds will reuse
the same server (for example when a test changes), and the
test resources server is shutdown at the end of the build
session.

Case 3: project local standalone mode

If the only task which is called is `startTestResourcesService`,
then a test resources server is started in the background and
will _outlive_ the build. There is an option to block the build
instead of running in the background.

When running in the background, the test server must be shutdown
by calling the `stopTestResourcesService` task.

Case 4: single build or continuous build connecting to a
standalone server

If a server is already started in standalone mode, instead of
starting a new test server, builds will reuse the test server
(unless a different _settings directory_ is set).

Case 5: shared test resources server

This is a variant of the standalone mode, where the settings
are written in a project-independent location, making it possible
to share a test resources server between multiple builds, in
multiple projects.