Add custom `testResources` source set
This commit adds a `testResources` source set when applying the
test resources plugin. This makes it easy to implement custom
test resources in a single project, by implementing the test
resource resolver in `src/testResources/java` for example.

The test resources declares in that source set are automatically
added to the test resources classpath.