Introduce support for multi-project builds
This commit introduces a way to provide a test resources service
which is shareable between projects of a multi-project build.
Without this, each project of a multi-project build applying the
test resources plugin would start a test resources service
independently of the others, which means that if 2 projects need
the same containers types, they wouldn't be able to re-use the
same container (unless they use the `shared` mode and make sure
that each project declares all test resolvers which are required
by all projects).

With this change, it is now possible to declare a project which
is the "test resources supplier". This project is responsible for
declaring all resources it can serve. Then each project willing
to use this test resources server need to apply a _different_
plugin, `io.micronaut.test-resources-consumer` and declare what
project provides the test resources server:

```gradle
micronaut {
   testResources.targetProject = ':testresources-supplier'
}
```

The configuration will then automatically be injected, and the
test resources service lifecycle will be handled appropriately.