Add custom `testResources` source set (#484)
* Add custom `testResources` source set

This commit adds a `testResources` source set when applying the
test resources plugin. This makes it easy to implement custom
test resources in a single project, by implementing the test
resource resolver in `src/testResources/java` for example.

The test resources declares in that source set are automatically
added to the test resources classpath.

* Introduce internal start task

This fixes the logic for determining if the server should be kept alive after
the build stops: now there's an internal start service task, which effectively
starts the server. We determine if the server should be kept alive by looking
if the task graph contains the _explicit_ task `startTestResourcesService`.