Introduce support for multi-project builds (#485)
* Add custom `testResources` source set

This commit adds a `testResources` source set when applying the
test resources plugin. This makes it easy to implement custom
test resources in a single project, by implementing the test
resource resolver in `src/testResources/java` for example.

The test resources declares in that source set are automatically
added to the test resources classpath.

* Introduce internal start task

This fixes the logic for determining if the server should be kept alive after
the build stops: now there's an internal start service task, which effectively
starts the server. We determine if the server should be kept alive by looking
if the task graph contains the _explicit_ task `startTestResourcesService`.

* Introduce support for multi-project builds

This commit introduces a way to provide a test resources service
which is shareable between projects of a multi-project build.
Without this, each project of a multi-project build applying the
test resources plugin would start a test resources service
independently of the others, which means that if 2 projects need
the same containers types, they wouldn't be able to re-use the
same container (unless they use the `shared` mode and make sure
that each project declares all test resolvers which are required
by all projects).

With this change, it is now possible to declare a project which
is the "test resources supplier". This project is responsible for
declaring all resources it can serve. Then each project willing
to use this test resources server need to apply a _different_
plugin, `io.micronaut.test-resources-consumer` and declare what
project provides the test resources server:

```gradle
micronaut {
   testResources.targetProject = ':testresources-supplier'
}
```

The configuration will then automatically be injected, and the
test resources service lifecycle will be handled appropriately.

* Disable tests on Windows since Docker is not available on CI

* Clarify error message

* Disable sonar

There is apparently no way we can make the build pass, it's too flaky with GraalVM
builds.