Integrate with native and AOT (#490)
* Integrate with native and AOT

This commit integrates the test resources support with the AOT
support and GraalVM native support, so that we can run:

- optimizedRun: the standard JVM AOT-optimized run
- nativeTest: the native version of tests
- nativeOptimizedRun: the native run

all with test resources support enabled. It's worth noting
that native run support is a bit hackish, in the sense that
if you run `nativeCompile` only, then it will produce a
native binary which is for production and doesn't contain
the test resources client, as expected.

However, if you run `nativeRun`, then the test resources
client (and the properties file containing the connection
information) need to be in the binary, so `nativeCompile`
would _include_ them. This is specific of native, because
in the JVM mode, we can simply add elements to the run
classpath. For native, there's no classpath so everything
needs to be in the binary. The consequence is that if
the user picks the result of the `nativeCompile` task
_after_ running `nativeRun`, then it would contain a
"wrong" binary containing the test resources client. This
is unlikely to be a problem in practice, though, since
you would likely not create a production artifact with
`nativeRun` before, since it would block the build.

In addition, this commit fixes an issue with the test
resources client classpath, which could include a different
version of Micronaut than the one picked by the user.

The reason is that the `developementOnly` configuration,
and similarly the configuration we create for resolving
the client, is _added_ to the run task classpath: it's
not a single dependency graph but 2 appended dependency
graphs. Therefore, this commit makes sure to override
the version used by the client to the one used by the
application.

* Fix micronaut version resolution