Fix CRaC for ubuntu 22.04
Previously, the CRaC JDK fils if using Ubuntu 22.04

After checking with the Azul team, this issue only affects Mac and WSL docker clients where the
kernel has not yet been updated to support it.

This fix sets and environment variable that has no effect for previous versions, but fixes the
kernel settings for 22.04.

I have made the default 22.04, and updated the docs accordingly