Fix custom dockerfile usage
If the user was using a custom Dockerfile (at the root of the project),
then the file would be used directly, but wouldn't be found by the
docker engine because living outside of the build context. This commit
fixes the problem so that the file is appropriately templated and
used in the build context.

Fixes #583