Fix `test` task being out of date with test resources
This commit upgrades to Micronaut Test Resources 1.1.5, which offers
the ability to pass the server connection settings to the client via
system properties instead of a file on classpath.

This fixes the problem of the `test` task being constantly out-of-date
when the `test-resources` plugin is applied. This does **not** fix
the issue with `nativeTest`, which is still out-of-date, but will
require a new release of the native build tools.

Fixes #579