Fix integration with Micronaut Test Resources
This commit fixes how the Gradle plugin integrates with Micronaut
Test Resources. The latest version of test resources makes use of
a fat jar, which allows us to inject the Micronaut Platform BOM
to the test resources service classpath, so that the inferred
dependencies, like the database drivers, can be resolved even if
the user didn't specify a particular version.

Previously to this change, it was possible that the database
drivers were accidentally resolved because the test resources
service itself was built with Micronaut, and therefore adding
the Micronaut core BOM on classpath.