Enable GraalVM metadata repository (#638)
* Enable GraalVM metadata repository

This commit enables use of the GraalVM metadata repository by default.
It changes how the files are copied when working on Docker images,
because previously, we had a single directory of resources (the
generated one), but now it is possible to have multiple directories.
For example, if netty is present on classpath, we will need to copy
the configuration files for netty.

In practice, when we prepare the docker build context, we need to
scan those directories and copy them. It is however possible that
several directories have the same name, because the directory names
are derived from GAV coordinates, so the same version can appear
multiple times.

In order to:
   - avoid overwriting configuration files for several modules
which have the same version
   - make things easier to debug when reading the docker files

we add inference to figure out if a configuration directory
comes from the metadata repository, in which case we wouldn't
use only the directory name, but the 3 last directories
which correspond to the GAV coordinates. For example,
instead of a directory named `1.0.0`, we would have a directory
named `org.company:artifact:1.0.0`.

* Fix Jakarta namespace

* Fix test under regular JVM