Fix regression in test framework usage (#644)
The fix introduced in #640 fixed the problem for JUnit as a test
framework, which is the default when using starter. However, it
broke when using Spock or Kotest.

However, all of the test frameworks that we use require JUnit
Platform. This means that the code was probably used when we
supported JUnit 4 without the legacy runner, and that it can
be simplified now. Basically as soon as we use a test framework
we can configure JUnit Platform.

Fixes #643