Fix integration with GraalVM repository (#654)
* Fix integration with GraalVM repository

This commit fixes a number of issues with building native images
which surfaced after the latest Micronaut snapshots removed
metadata from jar files. In particular:

- when building a native image under Docker, the "exclude-config"
flags were not propagated, which means that even if the metadata
repository was enabled, the configuration files embedded in jars
were still read when they shouldn't
- when building an AOT native image, the metadata repository was
ignored, as well as configuration excludes

* Remove GraalVM dev build

This is causing too much troubles, we will need to rework how
we pre-emptively test GraalVM builds.