Introduce option to customize generated docker files
This commit introduces the possibility to customize the dockerfiles
which are generated before building the docker image. To do this,
the user can either configure a single task, e.g:

```gradle
tasks.named("dockerfile") {
   ...
}
```

or all tasks at once:

```gradle
tasks.withType(DockerBuildOptions).configureEach {
   ...
}
```

Then a DSL is exposed which allows easily modifying
the generated file, for example:

```gradle
editDockerfile {
    after('COPY layers/libs /home/app/libs') {
        insert('COPY server.iprof /home/app/server.iprof')
    }
}
```

The editor DSL allows narrowing down what part of the
file to edit thanks to nested `before` and `after` calls.

Note that this is a "poor man's" solution to the problem
of customizing generated files, because using the current
`Dockerfile` task APIs it is extremely difficult or even
impossible to alter the configuration after it's done.

The editor DSL is minimal for now, in the sense that it
only supports strings, not patters, for simplicity.

Fixes #667