Copy the customized CRaC dockerfiles if they exist in the root
Previously, we allowed users to create `Dockerfile` and/or `DockerfileCracCheckpoint` in the root of their project and we then would attempt to use these for the image generation.

This doesn't work.  The Docker plugin requires all Docker content to be in the same file tree.

This PR copies checks for the existence of these files, and copies them to the correct location instead of building the dockerfiles.

Fixes #678