Fix duplicate entries on classpath
This commit significantly reworks how configurations are setup, in order to
fix a problem which appears independently in 2 locations: if using test
resources or the developmentOnly configuration, it was possible that
dependencies were duplicated, in different versions, on classpath.

For example, the test resources client was previously resolved independently
into its own dependency graph, which was then added to the run task
classpath. Similarly, the `developmentOnly` configuration was resolved
independently and added to the runtime classpath.

The problem is that by appending to the classpath, we have 2 independent
dependency graphs, which _may_, or may not, include the same dependencies.
In practice, it was easy to get a mismatch. For example, even if we
made sure to use the same Micronaut version for the client than the one
the user declares, it was possible for transitive dependencies like
Netty to have different versions.

The solution is to always use _single_ dependency graphs for the
runtime/build/native image/... classpaths. Then Gradle will properly
perform dependency conflict resolution as expected, and we won't
get any duplicate.

In addition, we changed how the `nativeRun` (and `nativeCompile`)
task interact with test resources: the previous implementation didn't
properly detect if the client should be injected or not.